Added .env for download location and improved download

+++ src/main/java/de/one1on/sgjscraper/core/Downloader.java
@@ -1,8 +1,14 @@
 package de.one1on.sgjscraper.core;
 
+import io.github.cdimascio.dotenv.Dotenv;
 import org.apache.commons.io.FileUtils;
+import org.apache.commons.io.FilenameUtils;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
-import java.io.*;
+import java.io.File;
+import java.io.IOException;
+import java.net.URISyntaxException;
 import java.net.URL;
 import java.util.List;
 import java.util.concurrent.CompletableFuture;
@@ -10,12 +16,35 @@ import java.util.stream.Collectors;
 
 public class Downloader {
 
+    public static Logger logger = LoggerFactory.getLogger(Downloader.class);
     private final List<String> hashtags;
     private List<String> urls;
+    private Dotenv env = Dotenv.load();
+    private File baseDir;
 
     public Downloader(List<String> hashtags, List<String> urls) {
         this.hashtags = hashtags;
         this.urls = urls;
+        initBaseDir();
+    }
+
+    private void initBaseDir() {
+        final String baseDir = env.get("IMG_LOCATION");
+        if(null == baseDir || baseDir.isEmpty()) {
+            this.baseDir = new File("build/images");
+            logger.warn("Using default image location {}", this.baseDir.getPath());
+        } else {
+            this.baseDir = new File(baseDir);
+            logger.info("Using image location {}", this.baseDir.getPath());
+        }
+
+        if(this.baseDir.mkdir()) {
+            logger.info("Created directory image location {}", this.baseDir.getPath());
+        }
+        if(!this.baseDir.canWrite()) {
+            throw new IllegalStateException("Can not write to directory " + this.baseDir.getAbsolutePath());
+        }
+        System.out.println();
     }
 
     public List<CompletableFuture<File>> download() {
@@ -31,19 +60,17 @@ public class Downloader {
         return CompletableFuture.supplyAsync(() -> {
             try {
                 return downloadFileA(url);
-            } catch (IOException e) {
+            } catch (IOException | URISyntaxException e) {
                 e.printStackTrace();
             }
             return null;
         });
     }
 
-     public File downloadFileA(String imageUrl) throws IOException {
+     public File downloadFileA(String imageUrl) throws IOException, URISyntaxException {
         URL url = new URL(imageUrl);
-        String fileName = url.getFile();
+        String fileName = FilenameUtils.getName(url.toURI().getPath());
         String destName = "figure-" + fileName.hashCode() + ".jpg";
-        final File imageDir = new File("build/images");
-        imageDir.mkdir();
 
          File file = null;
          if(hashtags.isEmpty()) {
@@ -51,12 +78,15 @@ public class Downloader {
          }
 
          for (String hashtag : hashtags) {
-             final File hashTagDir = new File(imageDir, hashtag);
+             final File hashTagDir = new File(baseDir, hashtag);
              hashTagDir.mkdir();
              file = new File(hashTagDir, destName);
-             System.out.println(file.getName());
+             logger.info("Image for hashtag {} - {}", hashtag, fileName);
+             if(!file.exists()) {
+                 logger.info(file.getName());
 
-             FileUtils.copyURLToFile(url, file);
+                 FileUtils.copyURLToFile(url, file);
+             }
          }
 
         return file;
